Some cleanups Simple XInput support in dev builds Project 2dfx 2.0 Lotta supported languages viewwindow in progress of tweaking
git-svn-id: svn://localhost/vcspc/trunk@38 9f344024-d43e-4b27-bde7-3d027e20192d

+++ VCS PC/WidescreenSupport.cpp
@@ -140,7 +140,7 @@ void WidescreenSupport::Recalculate(int nWidth, int nHeight, bool bAlways)
 		//fAimpointFix = (9.0f/12800.0f)*fNewResolutionWidthMultiplier + 0.1f;
 		fMapZonePosX2 = 7.5f/fNewResolutionWidthMultiplier;
 
-		fAimpointFix = 0.5f * atan(RAD_TO_DEG * SetAspectRatio()) /** ((4.0f/3.0f)/1.1f)*/;
+		fAimpointFix = 0.5f * static_cast<float>(atan(RAD_TO_DEG * SetAspectRatio())) /** ((4.0f/3.0f)/1.1f)*/;
 
 		f0pt7_h = 0.7f/fNewResolutionHeightMultiplier;
 		f0pt95_h = 0.95f/fNewResolutionHeightMultiplier;
